Reuse started build if rerunning CI build scripts
Useful when using these scripts to debug locally.

See e.g.: MethodicalAcceleratorDesign/MAD-X#1033.

+++ .github/build/manylinux1/madx.sh
@@ -5,23 +5,25 @@ set -ex
 # Must be run from the root the directory of the MAD-X sources.
 # Builds in './build' and installs to './dist'.
 
-rm -rf build
-mkdir build
+mkdir -p build
 cd build
 
 : ${PY:=/opt/python/cp36-cp36m/bin}
 
-$PY/pip install --upgrade cmake
-$PY/cmake .. \
-    -DBUILD_SHARED_LIBS=OFF \
-    -DMADX_STATIC=ON \
-    -DCMAKE_INSTALL_PREFIX=../dist \
-    -DCMAKE_BUILD_TYPE=Release \
-    -DCMAKE_C_FLAGS="-fvisibility=hidden" \
-    -DCMAKE_CXX_FLAGS="-fvisibility=hidden" \
-    -DCMAKE_Fortran_FLAGS="-fvisibility=hidden" \
-    -DMADX_INSTALL_DOC=OFF \
-    -DMADX_ONLINE=OFF \
-    -DMADX_FORCE_32=OFF \
-    -DMADX_X11=OFF
+if [[ ! -f CMakeCache.txt ]]; then
+    $PY/pip install --upgrade cmake
+    $PY/cmake .. \
+        -DBUILD_SHARED_LIBS=OFF \
+        -DMADX_STATIC=ON \
+        -DCMAKE_INSTALL_PREFIX=../dist \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DCMAKE_C_FLAGS="-fvisibility=hidden" \
+        -DCMAKE_CXX_FLAGS="-fvisibility=hidden" \
+        -DCMAKE_Fortran_FLAGS="-fvisibility=hidden" \
+        -DMADX_INSTALL_DOC=OFF \
+        -DMADX_ONLINE=OFF \
+        -DMADX_FORCE_32=OFF \
+        -DMADX_X11=OFF
+fi
+
 $PY/cmake --build . --target install -j
